LayoutTests/platform/ios-sim-deprecated/fast/files/file-reader-file-url-expected.txt

CONSOLE MESSAGE: line 9: Received data: 0123456789
Test that FileReader works under the file URL with file path separation enforced.
